backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv       # load the Gemini key from .env